Bypass auth when in the integration environment.
Summary:
github auth in the integration envioronment redirected us to an invalid url, and we needed to cut and past the correct github_oath code.
In this revision, we check the environment variable CLOUD_ENVIRONMENT, and bypass the auth steps if we are under integration environment.

It was suggested in https://twitchtv.atlassian.net/browse/SCIENG-307 that we can use a flag, checking the environment variable is more
straightforward as we use the same run script in prod and integration environments.

Test Plan:
go test ./...
integration test pass
manually test no more redirection in the integration environment

Reviewers: ojcm, mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3700